requirements.txt
fastapi == 0.95.1
PyJWT == 2.6.0
aiohttp == 3.8.4
aiosmtplib == 2.0.1
pycryptodome == 3.17
python-multipart == 0.0.5
haversine == 2.8.0
uvicorn == 0.20.0
jinja2 == 3.1.2
slowapi == 0.1.8
email_validator == 2.0.0.post2
python-magic == 0.4.27
